+++ test_all.sh
@@ -1,0 +1,4 @@
+# Run unit tests, and e2e tests.
+uv run pytest
+uv run pytest tests/e2e_tests -k full
+uv run pytest tests/e2e_tests -k concise