+++ CHANGELOG.md
@@ -17,6 +17,8 @@ These initial releases have usable behavior, but may have some rough edges for s
 #### Internal changes
 
 - Run tests in CI via tox.
+- Small helper script `test_all.sh` to run unit tests, then all e2e tests.
+- Benchmarking script to ensure consistent performance over time.
 
 ### 0.6.0